maint: add dev notes and update smoke test
* Add dedicated DEVELOPING.md for local dev
setup and notes, and reference that section in
the README. Also update links in README.
* Console output is needed for smoke test, but
DiagConsoleLogger doesn't seem to be logging as
expected, maybe because of the bundler? So for
smoke tests right now, use ConsoleSpanExporter.
* Sometimes the console output takes longer than
the default timeout of 4000ms, so update Cypress
config to timeout at 10000ms instead.