maint: add dependabot and unify otel dependencies (#43)
## Which problem is this PR solving?

- Closes #41 

## Short description of the changes

- maint: add dependabot (example deps update together, otel deps update
together, dev deps update together to reduce number of dependabot PRs
and also to avoid versioning conflicts)
- constrain versioning to avoid conflicts (with some packages being
stable and some experimental, there can be breaking changes and
conflicts if we're not careful with versioning; recommendation has been
to use tilde or pin versions so I figured we'd start with tilde and see
how it goes)
- unify otel dependencies (1.7 api with 1.19 stable with 0.46
experimental)
- update codeowner pipeline-team (previously named pipeline)

## How to verify that this has the expected result

dependabot starts opening PRs to update dependencies, keeping in groups
as defined. also minimal conflicts arise when installing in a new app.