feat: ✨ Browser attributes span processor (but also so much more) (#45)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
This PR adds the `BrowserAttributesSpanProcessor` which adds rich
browser attributes to every span. Attributes like the page location and
screen height and width might change during the lifecycle of a session
so they are better captured through a Span Processor rather than on the
Resource.

- Closes #11 

## Short description of the changes
- `BrowserAttributesSpanProcessor` class that adds attributes to each
span
- `CompositeSpanProcessor` class that combines multiple span processors
together since the base SDK does not support multiple span processors
- Span processor builder that configures the `BatchSpanProcessor` with
the exporter as well as the `BrowserAttributesSpanProcessor` and an
optional user provided span processor
- Had to switch the exporter being configured through this span
processor builder since the base SDK does not support configuring an
exporter through the `traceExporter` option if a `spanProcessor` is
provided.

## How to verify that this has the expected result
- Run the test app and check for the additional browser attributes
- The data should still be exported to Honeycomb

---------

Co-authored-by: Jamie Danielson <jamiedanielson@honeycomb.io>