feat: Auto-enable web vitals instrumentation (#106)
## Which problem is this PR solving?
- Closes #100 

Automatically includes `WebVitalsInstrumentation` in array of
instrumentations when SDK is initialized, unless the config option is
set explicitly to `false`. This is so that when the SDK is configured,
users get web vitals spans without having to configure them explicitly.

## Short description of the changes
- Adds `WebVitalsInstrumentation` as a default instrumentation in the
SDK
- Adds support for configuring `WebVitalsInstrumentation`, including
turning it off and supporting the other config values.
- Adds smoke tests for web vitals spans

**Note**: one interesting conflict that was surfaced by smoke tests
(shout out to smoke tests, so glad we have them) is that when the web
vitals instrumentation is initialized and includes sending spans for
`INP`, there is an [event listener added by the base `web-vitals`
library](https://github.com/GoogleChrome/web-vitals/blob/8058a007c9878c2b56c787040d4e8ca282dc203a/test/views/inp.njk#L85).
This means that if a user also adds an event listener for click or other
events, they will be collected twice. We should make a note of this in
our documentation. For now I removed the event listener in the example
to make sure the click is only collected once.

## How to verify that this has the expected result
- Run the example and see web vitals spans in the console
- Set the config to false and verify that no web vitals spans are being
sent


Co-authored-by: Wolfgang Therrien
<wolfgangcodes@users.noreply.github.com>

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1206754510479286

---------

Co-authored-by: Wolfgang Therrien <wolfgangcodes@users.noreply.github.com>