docs: Bump Readme to Beta (#116)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Description in Readme claimed the instrumentation is in Alpha. But we're
in Beta!

## Short description of the changes

## How to verify that this has the expected result
Read the readme, and confirm it does not say Alpha.