docs: README.md update (#118)
## Which problem is this PR solving?

An outdated link to important information about how to find your
Honeycomb API Key.

## Short description of the changes

This PR updates an outdated docs.honeycomb.io link to the correct
location. In addition, numbers are added to the steps in the Getting
started section.

## How to verify that this has the expected result

Does clicking on the docs link get you to the desired location?