feat: add browser.name attribute (#130)
## Which problem is this PR solving?
Adds a `browser.name` attribute to events

## Short description of the changes
Delegates to [ua-parser-js](https://www.npmjs.com/package/ua-parser-js)
so the list of available browsers comes from there. At its core, this is
just user-agent parsing, with all of the caveats that come with that:
it's fragile, a best guess effort, etc.

ex. I tested with [Arc](https://arc.net) and it's getting parsed as
Chrome; this is expected since Arc doesn't change the User Agent at all.

## How to verify that this has the expected result
Run the app and see `browser.name` attached: 

![image](https://github.com/honeycombio/honeycomb-opentelemetry-web/assets/11722214/3ab5c6a9-6a3e-4104-9609-2e66cf2e5315)