feat: Upgrade to web-vitals@4.0.0 (#152)
## Which problem is this PR solving?
Web vitals but released 4.0.0 which includes an updates to field names
and additional for `LCPAttribution`, `INPAttribution` and
`TTFBAttribution` and deprecated `onFID`. See the full changes
[here](https://github.com/GoogleChrome/web-vitals/blob/v4.0.0/docs/upgrading-to-v4.md).

## Short description of the changes
Upgrade to `web-vitals@4.0.0`. Full change log of the underlying changes
can be found
[here](https://github.com/GoogleChrome/web-vitals/blob/v4.0.0/docs/upgrading-to-v4.md).

For this release, we'll send both the updated field names and the
deprecated field names, which will be removed in the next major release.
Additionally, `onFID` has been deprecated and will removed in the next
major release.

For this instrumentation we have updated attribute payloads as follows:

### `LCP`:
- Updated `resource_load_time` -> `resource_load_duration`
- Deprecated `resource_load_time`

### `INP`:
- Updated `element` -> `interaction_target`
- Deprecated `element`
- Updated `event_type` -> `interaction_type`
- Deprecated `event_type`
- Added `interaction_time` 
- Added `input_delay`
- Added `interaction_target_element`
- Added `next_paint_time`
- Added `presentation_delay`
- Added `processing_duration`

### `TTFB`:
- Updated `waiting_time` -> `waiting_duration`
- Deprecated `waiting_time`
- Updated `dns_time` -> `dns_duration`
- Deprecated `dns_time`
- Updated `connection_time` -> `connection_duration`
- Deprecated `connection_time`
- Updated `request_time` -> `request_duration`
- Deprecated `request_time`
- Added `cache_duration`

## How to verify that this has the expected result
See that the new new and updated properties are being reported.