refactor: export multiple packages (#203)
## Which problem is this PR solving?
This PR lets us publish multiple packages from this repo

## Short description of the changes
This PR moves all of the main code to a
`packages/honeycomb-opentelemetry-web/` folder, and sets up the barest
scaffolding for a `packages/honeycomb-react-user-instrumentation/`
package.

This also adjusts the existing CircleCI configuration to only run on
changed files. Specifically, the main `.circleci/config.yml` now just
determines which packages have changed, and then triggers the
workflow(s) in `.circleci/build.yml` with the correct config options.

`.circle/build.yml` contains TWO workflows: one for the web distro and
the other for the react instrumentation. Each workflow will only get
triggered if files in that directory have changed.

I also updated `RELEASING.md` with more details on our new release
process. I am not sure how to test this without merging (is that even
possible?) so I anticipate we may have to revisit those guidelines in
the future, but for now reviewers should weigh in on whether we think
the process described in that document is practical.

## How to verify that this has the expected result
I used [this
pr](https://github.com/honeycombio/honeycomb-opentelemetry-web/pull/204)
to test these changes, and the build/test pipeline there is currently
green.

This _should_ work for the release steps as well, but I have no way of
testing that without merging this to main.