feat: support multiple exporters (#303)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
We have had multiple requests to add custom exporters (plural) rather
than just the one. This adds a field to the honeycomb SDK configuration
that allows providing an array of exporters.

- Closes #<enter issue here>

## Short description of the changes

This adds the `traceExporters` field which takes an array of exporters,
it works in the same way that the `traceExporter` field except this
takes an array of exporters rather than just the one.

## How to verify that this has the expected result

Smoke test is a WIP but we do have an example project
`hello-world-custom-exporter` that shows how to specify multiple
exporters including the default one.