feat: Allow Disabling the default Exporter (#328)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

We have had a few requests come in asking for the ability to disable the
default exporter. This makes it possible through a new property in the
configuration.

## Short description of the changes

This adds a new SDK configuration property `disableDefaultTraceExporter`
that, when set to true will disable the default exporter.

## How to verify that this has the expected result

Provide this boolean as `true` in an example and note no events are sent
to honeycomb.

---------

Co-authored-by: Wolfgang Therrien <wolfgangtherrien@honeycomb.io>