fix: This code duplicated  not needed. (#329)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Short description of the changes
This is the original implementation for adding data attributes. it was
not removed we updated to use the `dataset` API.

## How to verify that this has the expected result
Use the test apps to ensure that the data attributes are still being
attached to spans.