feat(error-autoinstrumentation): Add applyCustomAttribitesToSpan config option (#447)
## Which problem is this PR solving?
Generally auto-instrumentation should include a callback config option
that allows mutating the span (to add attributes / whatever else) which
was missing in our config. This PR adds the
`applyCustomAttribitesToSpan` that exposes the error object and the span
through a callback function allowing the user to mutate each error span
in the config. This option is also on the `recordException` function.
Ideally we will eventually deprecate the `attributes` argument on this
function but that would be a breaking change right now so I'll leave
that for a future PR.

## Short description of the changes
- Adds `applyCustomAttribitesToSpan` config option to error
instrumentation
- Adds `applyCustomAttribitesToSpan` argument to `recordException`
function
- Docs
- Tests

## How to verify that this has the expected result
- Tests pass