maint: Peer dependencies and testing for v2 network instrumentation (#493)
## Which problem is this PR solving?
The latest version of `@opentelemetry/auto-instrumentations-web` has
dependencies on `v0.200.0` of `@opentelemetry/instrumentation-fetch` and
`@opentelemetry/instrumentation-xml-http-request`. This is a major
version bump and we need to do some testing to ensure that the SDK is
compatible with the latest versions of these instrumentations.

## Short description of the changes
- Added a peer dependency for `@opentelemetry/auto-instrumentations-web`
and specified the compatible version.
- The latest version of `@opentelemetry/context-zone` is also a major
version bump to `v2.0.0`. Included a peer dependency for it as well.
- Upgraded examples to use the latest versions of peer dependencies.
- Added smoke tests to test XHR and fetch instrumentation.

## How to verify that this has the expected result
- [x] Smoke tests pass.
- [x] Example apps run.
- [x] Spans from `@opentelemetry/instrumentation-fetch` and
`@opentelemetry/instrumentation-xml-http-request` are sending to
Honeycomb.
- [x] `@opentelemetry/context-zone` nests spans appropriately.