maint: Upgrade to v2.0.0/v0.200.0 upstream OTel packages (#503)
## Which problem is this PR solving?
- Closes #501 

Internal refactoring to upgrade dependencies to v2.0.0/v0.200.0 upstream
OTel packages. There should be **no changes** to the external facing
API.

Relied heavily on the [upgrade
guide](https://github.com/open-telemetry/opentelemetry-js/blob/main/doc/upgrade-to-2.x.md#-opentelemetryresources-changes-for-implementors-of-resource-detectors)
to make the changes.

As a bonus, multiple span processors are supported by default in v2 of
the `WebTracerProvider` so we got to delete a lot of the code we had to
make that happen.

## Short description of the changes
- Resource APIs were one of the most affected, had to make API and type
changes to all of this.
- `Span` API is no longer exposed so had to move a lot of tests to
creating a span with the tracer and inspecting it through an in memory
exporter.
- Got rid of `CompositeSpanProcessor` since `WebTracerProvider` supports
an array of span processors now.

## How to verify that this has the expected result
- Unit tests pass
- Smoke tests pass
- Manual testing

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>