feat: add an option to disable automatic browser attributes (#510)
## Which problem is this PR solving?

It is currently not possible to disable automatic attributes inferred by
the browser, which is inconsistent with the options to turn off other
auto-instrumentation. It also prevents this SDK from being used in
non-browser web contexts.

## Short description of the changes

Adds a new `disableBrowserAttributes` config option, which defaults to
false. Turning on this option prevents the Honeycomb SDK from
automatically adding browser-derived resource attributes, as well as the
browser attribute span processor.

## How to verify that this has the expected result

- New unit tests added.
- Smoke tests still pass.