chore: Upgrade to web-vitals 5.0.0 (#534)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes
https://github.com/honeycombio/honeycomb-opentelemetry-web/issues/531

## Short description of the changes
Upgrades web-vitals to `5.0.0` this removes `FID` from the CWV since it
was replaced by INP.

The full change log details are here:

https://github.com/GoogleChrome/web-vitals/blob/main/docs/upgrading-to-v5.md
and 

https://github.com/GoogleChrome/web-vitals/blob/main/CHANGELOG.md#v500-2025-05-07