fix: honoring custom sampler when provided by user. (#559)
## Which problem is this PR solving?

- Closes #557 

## Short description of the changes
This change provides a small quick fix to:
- return custom sampler in case user provided it through the options
when running HNY SDK
- renaming the function configureDeterministicSampler to
configureSampler, since this does not only configure deterministic
sampler (function name a bit misleading)
- adding unit test to test the functionality.

## How to verify that this has the expected result
provided custom sampler in option will be used when provided. Internal
deterministic sampler is used when the custom sampler does not exist,
and sample rate is provided.