fix: remove redundant call to setGlobalTracerProvider as its set upon creation. (#615)
… creation.

<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes
https://github.com/honeycombio/honeycomb-opentelemetry-web/issues/612

## Short description of the changes
`setGlobalTracerProvider` is called as a side effect of `new
WebTracerProvider(...)` we don't need to do this manulally.

## How to verify that this has the expected result
See that the console error is no longer appearing.