feat: add page visibility state as a browser attribute (#624)
## Which problem is this PR solving?
While investigating a separate issue with LCP metrics appearing
inconsistent, we proposed the theory that pages are getting opened in
the background, and realized that having page visibility data will be
useful in general.

## Short description of the changes
Adds a new `browser.page.visibility` attribute that maps to [the
`document.visiblityState`
API](https://developer.mozilla.org/en-US/docs/Web/API/Document/visibilityState).

## How to verify that this has the expected result
I tried writing a smoke test for this and didn't see any values appear,
which I assume is because `document.visibilityState` doesn't apply in a
headless browser 🙃

I ran the examples locally and saw the attributes in the spans that got
emitted.

I updated the unit tests as well so at least we have that.